requirements.txt
Django==1.9
requests
django-filter
django-sendfile
djangorestframework
galaxy-lib
semver
